refactor(npmrc): extract inline test modules into dedicated files
Move three inline `#[cfg(test)] mod tests { ... }` blocks into
external `tests.rs` files following CODE_STYLE_GUIDE.md:
`workspace_yaml` (159 lines), `custom_deserializer` (72),
`npmrc_auth` (71). No behavioral change.

Part of #358.
